Move tooltip helper to Message utils while adding...
types.
